Fixed bug in low and high pass cortical filters.
This bug applied to both the matlab and julia implementations.
The matlab implementation is sensitive to rate order
(and assumes the rates and scales will be ordered highest to lowest). The julia
implementation got the high and low pass rate filters wrong, because it didn't
account for the negative rates in determining when to use high and low pass
filters.